refactored code, created dedicated entrypoints for e2e/ct modes
- removed dead code
