+++ packages/server/lib/modes/index.js
@@ -1,15 +1,23 @@
 module.exports = (mode, options) => {
+  const { testingType } = options
+
   switch (mode) {
-    case 'componentTestingInteractive':
-      return require('./component-testing').launchInteractiveMode(options)
-    case 'componentTestingRun':
-      return require('./component-testing').launchRunMode(options)
     case 'record':
       return require('./record').run(options)
-    case 'e2eRun':
-      return require('./run').run(options)
-    case 'e2eInteractive':
-      return require('./interactive').run(options)
+    case 'run':
+      switch (testingType) {
+        case 'e2e':
+          return require('./run-e2e').run(options)
+        case 'component':
+          return require('./run-ct').run(options)
+      }
+    case 'interactive':
+      switch (testingType) {
+        case 'e2e':
+          return require('./interactive-e2e').run(options)
+        case 'component':
+          return require('./interactive-ct').run(options)
+      }
     case 'smokeTest':
       return require('./smoke_test').run(options)
     default:
