+++ packages/server/lib/modes/interactive-ct.js
@@ -1,0 +1,13 @@
+const serverCt = require('@packages/server-ct')
+
+const run = (options) => {
+  const { projectRoot } = options
+
+  options.browser = options.browser || 'chrome'
+
+  return serverCt.start(projectRoot, options)
+}
+
+module.exports = {
+  run
+}
