+++ packages/server/lib/modes/interactive-e2e.js
