+++ packages/server/lib/modes/run-e2e.js
@@ -1,0 +1,10 @@
+const run = (options) => {
+  // if we're in run mode with component
+  // testing then just pass this through
+  // without waiting on electron to be ready
+  return require('./run').run(options)
+}
+
+module.exports = {
+  run
+}
